requirements.txt
matplotlib==3.9.3
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.5.2
seaborn==0.13.2
torchinfo==1.8.0
tqdm==4.67.1
timm